pj25/requirements.txt
youtube-transcript-api>=1.1.1
openai>=1.0.0
python-dotenv>=1.0.0
